changed ISBN_number model field to big int field

+++ list_of_books/models.py
@@ -5,7 +5,7 @@ class Book(models.Model):
     title = models.CharField(max_length=200)
     authors = models.CharField(max_length=200)
     published_date = models.CharField(max_length=20, null=True, blank=True)
-    ISBN_number = models.IntegerField(null=True, blank=True)
+    ISBN_number = models.BigIntegerField(null=True, blank=True)
     page_count = models.IntegerField(null=True, blank=True)
     thumbnail = models.CharField(max_length=255, blank=True)
     language = models.CharField(max_length=5, null=True, blank=True)
